LayoutTests/imported/w3c/web-platform-tests/html/semantics/forms/the-selectmenu-element/selectmenu-shadow-root-replacement.tentative-expected.txt


FAIL Test that the UA provided shadow root can be replaced promise_test: Unhandled rejection with value: object "NotSupportedError: The operation is not supported."
FAIL Test that the UA provided shadow root can be replaced using declarative shadow DOM assert_equals: expected (boolean) false but got (undefined) undefined
FAIL Test that only the UA provided shadow root can be replaced The operation is not supported.
